removed old edit field
