--- src/widgets/edit-parameter/EditParameter.scss
@@ -1,1 +1,0 @@
-
